SkyReplacement/requirements.txt
addict==2.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
Cython==0.29.30
docker-pycreds==0.4.0
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mmcls==0.23.0
mmcv-full==1.5.1
mmsegmentation==0.24.1
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.0.1
prettytable==3.3.0
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pycocotools==2.0.4
pykuwahara==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
torch==1.7.0
torchaudio==0.7.0a0+ac17b64
torchvision==0.8.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.9
wandb==0.12.17
wcwidth==0.2.5
yapf==0.32.0
